requirements.txt
# Core Framework
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.3

# Database
pymongo==4.8.0
dnspython==2.6.1

# AI / Gemini
google-generativeai==0.8.3
Pillow==10.4.0
pdf2image==1.17.0
PyMuPDF==1.24.10

# Data Processing
pandas==2.2.2
openpyxl==3.1.5
python-dateutil==2.9.0
pycountry==24.6.1
babel==2.15.0

# Utilities
python-dotenv==1.0.1
python-magic==0.4.27
shortuuid==1.0.13
bcrypt==4.2.0

# File Handling
boto3==1.35.3

# Validation
marshmallow==3.22.0
email-validator==2.2.0

# HTTP & Async
requests==2.32.3
gunicorn==23.0.0
